드랍율에 소수점 입력안되는 버그 수정함

+++ src/NewItem.tsx
@@ -58,7 +58,7 @@ const NewItem: React.FC<Props> = ({ onAddItem }) => {
   const handleAddClick = () => {
     onAddItem({
       name: name,
-      dropRate: parseInt(percent) / 100,
+      dropRate: parseFloat(percent) / 100,
       trials: 0,
       trialsPerClick: "1",
       createdAt: new Date()
